Create NBA_API.py
we will use the NBA API to determine how well the Golden State Warriors performed against the Toronto Raptors. We will use the API to determine the number of points the Golden State Warriors won or lost by for each game. So if the value is three, the Golden State Warriors won by three points. Similarly it the Golden State Warriors lost by two points the result will be negative two. 


We can plot out the PLUS MINUS column for the dataframes games_home and  games_away. We see the warriors played better at home.